Echo posts in a pager
